ETL_Pipeline/project-churn/requirements.txt
pandas>=1.5
numpy
scikit-learn
xgboost
sqlalchemy
psycopg2-binary
prefect
joblib
shap
python-dotenv
pytest
alembic
pydantic
pandas-profiling
mlflow
pytest-cov
black
isort
bandit
safety
great-expectations
jupyterlab
plotly
pandas
numpy
scikit-learn
xgboost
streamlit
plotly
seaborn
matplotlib
prefect
joblib
